Add notFourK to mixins; viewport testing unsucessful
There's a new place and a picture to explain it - three viewports. The first, is-hidden-desktop will show only for mobile and tablets. The second is the new one; something like is-hidden-touchand4k will display from desktop to quad hd. The final viewport, is-hidden-notfourk, will display the 4k content and styling. Currently, the page is rendering two sets of columns from is-hidden-touch and is-hidden-not4k, even at 4k. There are still issues.